requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
click==8.1.5
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
importlib-metadata==6.8.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
numpy==1.24.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.7.1
uuid==1.30
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2